requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.10.1
django-dotenv==1.4.2
django-filter==21.1
django-heroku==0.3.1
django-import-export==2.7.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.1.5
psycopg2==2.9.3
psycopg2-binary==2.8.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
typing-extensions==4.1.1
whitenoise==5.3.0
xlrd==1.2.0
xlwt==1.3.0
